+++ src/services/Fields.php
@@ -274,6 +274,7 @@ class Fields extends Component
                     /** @var Element[] $otherTargets */
                     $otherTargets = $owner::find()
                         ->drafts($owner->getIsDraft())
+                        ->provisionalDrafts($owner->getIsProvisionalDraft())
                         ->revisions($owner->getIsRevision())
                         ->id($owner->id)
                         ->siteId($otherSiteIds)
@@ -405,6 +406,7 @@ class Fields extends Component
                 /** @var Element[] $otherSources */
                 $otherSources = $target::find()
                     ->drafts($source->getIsDraft())
+                    ->provisionalDrafts($source->getIsProvisionalDraft())
                     ->revisions($source->getIsRevision())
                     ->id($source->id)
                     ->siteId($otherSiteIds)
@@ -413,6 +415,7 @@ class Fields extends Component
                 /** @var Element[] $otherTargets */
                 $otherTargets = $target::find()
                     ->drafts($target->getIsDraft())
+                    ->provisionalDrafts($source->getIsProvisionalDraft())
                     ->revisions($target->getIsRevision())
                     ->id($target->id)
                     ->siteId($otherSiteIds)